pkg/cli/cli.go shrinks from 1999 LOC to 418 LOC by relocating completer engine, peer-dial plumbing, sessionFilter, builtinApps, link/chrony/proto helpers, applyToDataplane/reloadSyslog, and the seven straggler handleShow* dispatcher methods into sibling files inside the same pkg/cli/ package. The residual cli.go is now a coherent entry-point file holding only the CLI struct, New, the 12 Set* methods, applyResult/dataplaneLoaded accessors, the Run() interactive loop, and the four prompt helpers.
Pure code motion — no signatures changed, no behavior changed, no
new init functions, no allocation rules altered. Public API surface
(cli.New + Set*/Run consumed by pkg/daemon/daemon_run.go) is
unchanged.

New sibling files:
- cli_show_security_dispatch.go: handleShowSecurity (250 LOC) +
  handleShowScreen + four security helpers (enabledStr,
  parsePolicyZoneFilter, resolveAddressDetail, printAppDetail).
  Dedicated file rather than appending 273 LOC to the already-1986
  LOC cli_show_security.go.
- peer.go: dialPeer + requestPeerSystemAction (cluster fabric gRPC).
- permissions.go: checkPermission (RBAC by login class).
- apply.go: applyToDataplane + reloadSyslog (CLI-side commit glue;
  skipped when daemon wires applyConfigFn).
- session_filter.go: sessionFilter type + parseSessionFilter +
  matcher methods + fetchPeerSessions/fetchPeerSessionSummary (the
  RPC fetchers are filter-bound and stay co-located).
- app_resolve.go: builtinApp/builtinApps + resolveAppName +
  resolveAddress.
- link.go: sysfs link-speed/duplex helpers.
- chrony.go: chronyc tracking output parser.
- proto.go: protoName{FromNum,ToID}, splitAddrPort, uint32ToIP,
  sessionStateName, ntohs, monotonicSeconds, capitalizeFirst.

Extensions to existing files:
- completion.go: cliCompleter type + Do method,
  completeConfigWithDesc, resolveCommand, formatAmbiguousMatches,
  completionNode alias, operationalTree/configTopLevel vars. Merged
  into the file that already held the completer's supporting helpers.
- cli_show_cluster.go: fabricRedirectCounters type +
  readFabricRedirectCounters (single consumer at :168).
- cli_show_flow.go: topTalkerEntry (consumers at :721/:744/:781).
- cli_show_nat.go: handleShowNAT dispatcher.
- cli_show_routing.go: handleShowRoute + handleShowProtocols.
- cli_show_services.go: handleShowClassOfService + handleShowServices.
- cli_show_interfaces.go: dhcpLease (only consumer is at :311/:322).

Plan + adversarial review

Plan doc: docs/pr/1444-cli-presenters/plan.md

- Codex round-1: PLAN-NEEDS-MAJOR (task task-mpmuqaqp-wdbuti)
- Gemini round-1: PLAN-NEEDS-MINOR (task task-mpmuqvdl-f976xd)
- Plan v2 addressed all v1 findings: drop redundant cli_completion.go,
  use bare names for new utility files, relocate
  readFabricRedirectCounters/topTalkerEntry to their consumer files,
  move security helpers to security dispatcher, create dedicated
  security dispatch file rather than bloating the 1986-LOC
  cli_show_security.go, classify dead code explicitly.
- Codex round-2 (3 attempts): sandbox infra failed twice; third
  retry with inline plan paste returned PLAN-NEEDS-MINOR with
  three ownership findings (task task-mpmvavek-dt62fr).
- Gemini round-2: PLAN-READY (task task-mpmv3mlw-4g2inm).
- Plan v3 acknowledges Codex r2 minor: applyResult stays in cli.go
  (core accessor with consumers across all show files), dhcpLease
  moves to cli_show_interfaces.go (only consumer), session_filter.go
  retains RPC fetchers (filter-bound, justified inline).

Retirement-boundary canary

Updated pkg/dataplane/retirement_boundary_canary_test.go allowlist
for 4 new pkg/cli files that retain legacy dataplane imports during
the #1373/#1451 migration:
- cli_show_cluster.go (fabric counters via dataplane.Telemetry)
- cli_show_security_dispatch.go (MaxRulesPerPolicy + counters)
- proto.go (SessState* enum + ProtoICMPv6 sentinel)
- session_filter.go (session key/value types + SessFlag*)

Mirror entries added to
docs/pr/1373-retire-ebpf-dataplane/README.md table to keep the
docs-mention canary in sync.

Test plan

- [x] go build ./... clean
- [x] go test ./pkg/cli/... pass
- [x] go test ./cmd/cli/... pass
- [x] go test ./... — all 32 test packages pass
- [x] sole external consumer (pkg/daemon/daemon_run.go) compiles
  unchanged
- [x] wc -l pkg/cli/cli.go = 418 (target was 600-700; came in lower
  because the residual is just struct/setters/Run/prompts)

No dataplane smoke required — pure code motion in Go control
plane, no dataplane touch. Wave-1 batch-merge model per user
guidance.

Closes #1444
