+++ docs/pr/1444-cli-presenters/reviewer-ids.md
@@ -1,0 +1,12 @@
+# #1444 cli presenters reviewer task IDs
+
+## Plan review round 1 (commit 69362886)
+
+- Codex: `task-mpmuqaqp-wdbuti` — PLAN-NEEDS-MAJOR
+- Gemini: `task-mpmuqvdl-f976xd` — PLAN-NEEDS-MINOR
+
+## Plan review round 2 (commit e7bf1668)
+
+- Codex: `task-mpmv39ez-5qk9gk` — NO-VERDICT (sandbox infra failed)
+- Codex retry: `task-mpmv8pqc-jrl9e1`
+- Gemini: `task-mpmv3mlw-4g2inm` — PLAN-READY
