+++ pkg/cli/cli_show_cluster.go
@@ -8,10 +8,41 @@ import (
 	"strings"
 
 	"github.com/psaab/xpf/pkg/cmdtree"
+	"github.com/psaab/xpf/pkg/dataplane"
 	dpuserspace "github.com/psaab/xpf/pkg/dataplane/userspace"
 	"github.com/vishvananda/netlink"
 	"golang.org/x/sys/unix"
 )
+
+// fabricRedirectCounters captures global counters describing how many
+// packets traversed the fabric redirect path.
+type fabricRedirectCounters struct {
+	total uint64
+	fab0  uint64
+	fab1  uint64
+	zone  uint64
+	drops uint64
+}
+
+// readFabricRedirectCounters samples the dataplane fabric counters.
+// Returns (zeroed, false) when the dataplane is not loaded.
+func (c *CLI) readFabricRedirectCounters() (fabricRedirectCounters, bool) {
+	if !c.dataplaneLoaded() {
+		return fabricRedirectCounters{}, false
+	}
+	telemetry := dataplane.TelemetryOf(c.dp)
+	read := func(index uint32) uint64 {
+		v, _ := telemetry.GlobalCounter(index)
+		return v
+	}
+	return fabricRedirectCounters{
+		total: read(dataplane.GlobalCtrFabricRedirect),
+		fab0:  read(dataplane.GlobalCtrFabricRedirectFab0),
+		fab1:  read(dataplane.GlobalCtrFabricRedirectFab1),
+		zone:  read(dataplane.GlobalCtrFabricRedirectZone),
+		drops: read(dataplane.GlobalCtrFabricFwdDrop),
+	}, true
+}
 
 // showChassis shows hardware information (like Junos "show chassis hardware").
