+++ pkg/dataplane/retirement_boundary_canary_test.go
@@ -42,9 +42,13 @@ var legacyDataplaneImportAllowlist = map[string]string{
 	"pkg/cli/cli.go":                           "embedded CLI constructor still stores the legacy bridge",
 	"pkg/cli/cli_clear.go":                     "clear commands still delete legacy session entries",
 	"pkg/cli/runtime.go":                       "#1517 cliRuntime interface declares the narrow CLI surface; still depends on root pkg/dataplane type names (SessionKey, CounterValue, etc.) until those types move to a domain package",
+	"pkg/cli/cli_show_cluster.go":              "#1444: fabric redirect counters relocated from cli.go; still reads legacy GlobalCtrFabric* counter indices via dataplane.Telemetry",
 	"pkg/cli/cli_show_flow.go":                 "flow display still uses legacy session keys and values",
 	"pkg/cli/cli_show_nat.go":                  "NAT display still uses legacy NAT/session metadata",
 	"pkg/cli/cli_show_security.go":             "security display still uses legacy counters and filter types",
+	"pkg/cli/cli_show_security_dispatch.go":   "#1444: handleShowSecurity dispatcher relocated from cli.go; still uses legacy MaxRulesPerPolicy + counter accessors",
+	"pkg/cli/proto.go":                         "#1444: shared session/proto helpers relocated from cli.go; still names dataplane.SessState* enum and ProtoICMPv6 sentinel",
+	"pkg/cli/session_filter.go":                "#1444: session filter type + RPC fetchers relocated from cli.go; still uses legacy session key/value types and SessFlag*",
 	"pkg/cluster/runtime.go":                   "clusterRuntime interface still names dataplane.SessionStore/Telemetry domain types from pkg/dataplane (#1518)",
 	"pkg/cluster/sync.go":                      "session sync still installs sessions through the legacy bridge",
 	"pkg/cluster/sync_bulk.go":                 "bulk sync still serializes legacy session entries",
